Design(Header): share 버튼 생성 - camera mode 버튼 Simulator로 이동
